+++ CHANGELOG.md
@@ -8,6 +8,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Fixed
 - Helpers were registered twice by mistake resulting in unecessary function calls.
 
+### Added
+- Added config objets to allow users to customise some helpers
+
 ## [0.4.1] - 2020-11-16
 ### Fixed
 - [$truncate]: Fixes an issue where the ellipsis would show even if the source length was shorter.